b/c we have two ActivityAdminControllers in the system: AllReady.Areas.Admin.Controllers.ActivityController AllReady.Controllers.ActivityController
also, the names of the controllers .cs files on disk do not always match the name of the controller class name which is even more confusing

b/c of this confusion, unit tests were added to the wrong unit test file for the wrong controller.

I'd recommend naming all controllers in the /Areas/Admin with the word "Admin" so we don't have any more mixups
